LFLT-597 Move translation management features into LPMC  - Extracted tooltip into a standalone component
