+++ src/components/users/ViewUserScreen.tsx
@@ -1,14 +1,14 @@
 import { CardWithTitle, PageOperationCard } from "@/components/common/Cards";
 import { DataRow, NarrowDataCell, WideDataCell } from "@/components/common/DataRow";
 import { MultiPaneScreen, NarrowPane, WidePane } from "@/components/common/ScreenLayout";
+import { Tooltip } from "@/components/common/Tooltip";
 import { PageOperationButton } from "@/components/navigation/OperationButton";
 import { Permission } from "@/core/domain/auth";
 import { Role, UserModel } from "@/core/model/user";
 import { dateFormatter } from "@/core/util/date-formatter";
 import { useSessionHelper } from "@/hooks/use-session-helper";
-import { faCircleInfo, faList, faUserGroup } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import React, { ReactNode, useEffect } from "react";
+import { faList, faUserGroup } from "@fortawesome/free-solid-svg-icons";
+import React, { ReactNode } from "react";
 import { useTranslation } from "react-i18next";
 
 interface ViewUserScreenParameters {
@@ -17,28 +17,9 @@ interface ViewUserScreenParameters {
 
 const ExternalUserTooltip = ({ user }: { user: UserModel }): ReactNode => {
 
-  const { t } = useTranslation();
-
-  useEffect(() => {
-    const init = async () => {
-      const { Tooltip, initTE } = await import("tw-elements");
-      initTE({ Tooltip }, { allowReinits: true });
-    };
-    init();
-  }, []);
-
-  if (user.role !== Role.EXTERNAL_USER) {
-    return null;
-  }
-
-  return (
-    <span
-      className="transititext-primary text-neutral transition duration-150 ease-in-out hover:text-neutral-600 focus:text-neutral-600 active:text-neutral-700 dark:text-neutral-400 dark:hover:text-neutral-500 dark:focus:text-neutral-500 dark:active:text-neutral-600"
-      data-te-toggle="tooltip"
-      title={t("user.label.external-user-role-locked")}>
-      <FontAwesomeIcon icon={faCircleInfo} />
-    </span>
-  )
+  return user.role === Role.EXTERNAL_USER
+    ? (<Tooltip tooltipKey={"user.label.external-user-role-locked"} />)
+    : null
 }
 
 /**
